Continue adding and split into files
