+++ embedded/app/include/bms_fault_checking.h
@@ -1,0 +1,6 @@
+#ifndef BMS_SM_H
+#define BMS_SM_H
+
+
+
+#endif
