+++ embedded/app/include/pressure_fault_checking.h
@@ -1,0 +1,7 @@
+#ifndef PRESSURE_SM_H
+#define PRESSURE_SM_H
+
+bool checkPrerunPressures(void);
+bool checkBrakingPressures(void);
+
+#endif
